+++ doc/sources/preview.rst
@@ -81,14 +81,14 @@ Current list of preview estimators:
    :align: left
 
    * - Estimator name
-     - Module
      - Is patching supported
+   * - :obj:`sklearn.linear_model.Lasso`
+     - Yes
+   * - :obj:`sklearn.covariance.ElasticNet`
+     - Yes
    * - :obj:`sklearn.covariance.EmpiricalCovariance`
-     - ``sklearnex.preview.covariance``
      - Yes
    * - :obj:`sklearn.decomposition.IncrementalPCA`
-     - ``sklearnex.preview.decomposition``
      - Yes
    * - :obj:`sklearn.linear_model.LogisticRegressionCV`
-     - ``sklearnex.preview.linear_model``
      - Yes
